requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-PyMongo==2.3.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
pymongo==4.10.1
pytest==8.3.4
python-dotenv==1.0.1
typing_extensions==4.12.2
Werkzeug==3.1.3
